Changing positioning and text
Changed fonts and the text positioning for the many content. Moved the may content from the bottom to the top so someone viewing the website wouldn't have to scroll as much (also changed some spacing so that it looks pretty). Changed the back ground to something that looks nice with the black text.